app/backend/requirements.txt
# Requirements for the REST API backend
flask==2.3.3
flask-pymongo==2.3.0
flask-cors==4.0.0
python-dotenv==1.0.0
pymongo==4.5.0
gunicorn==21.2.0

# Testing frameworks
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Mocking and fixtures
pytest-mock==3.12.0
mongomock==4.1.2
factory-boy==3.3.0
faker==19.12.0

# Coverage and reporting
coverage==7.3.2

# Additional test utilities
requests-mock==1.11.0

# Monitoring (for external Prometheus scraping)
prometheus-client==0.19.0